Moved function to locations dbaccessors
